app/requirements.txt
# ===============================
# LangChain ecosystem
# ===============================
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph

# ===============================
# LLM / Embedding
# ===============================
torch
transformers
sentence-transformers
scikit-learn
numpy
sentencepiece

# ===============================
# FastAPI & ASGI
# ===============================
fastapi
uvicorn[standard]
python-multipart

# ===============================
# Database
# ===============================
sqlalchemy
pyodbc

# ===============================
# Vector Database (NEW)
# ===============================
qdrant-client>=1.7.0

# ===============================
# Async Task Queue (NEW)
# ===============================
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0

# ===============================
# Google / Cloud
# ===============================
google-genai
google-cloud-storage

# ===============================
# NLP / Utils
# ===============================
langdetect
deep-translator
ddgs
chatterbox-tts>=0.1.6
torchaudio>=2.0.0
pillow

# ===============================
# Security
# ===============================
PyJWT
pwdlib[argon2]

# ===============================
# Infra / Utils
# ===============================
requests
python-dotenv
pydantic
